requirements.txt
# API clients
google-generativeai==0.3.2
requests==2.31.0
newsapi-python==0.2.7

# Utilities
python-dotenv==1.0.0

# Testing
pytest==7.4.3
pytest-mock==3.12.0

# Optional: for better JSON handling
pydantic==2.5.3
